Moved TestTaskExecutor to a new file
