+++ src/main/scala/higherkindness/rules_scala/workers/zinc/test/TestFrameworkRunner.scala
@@ -98,7 +98,6 @@ class ProcessTestRunner(
       }
     }
 
-    val taskExecutor = new TestTaskExecutor(logger)
     val failures = mutable.Set[String]()
     tests.foreach { test =>
       val process = new ProcessBuilder((command.executable +: command.arguments): _*)
